Share the shared-config-dir cleanup instead of hardening one class at a time
TokenStoreProfileTests' [Before(Test)] hook deleted the shared KCAP_CONFIG_DIR
artefacts with bare File.Delete / Directory.Delete. On Windows an open handle makes
that a hard sharing violation, and thrown from a Before hook it fails the test
before it runs:

  BeforeTestException: BeforeTest hook failed: The process cannot access the file
  '...\kcap-repopathstore-tests-952701da\config.json' because it is being used by
  another process.

A retrying cleanup for exactly this already existed — privately, inside
CrossProcessRefreshTests, from the previous round of this fix. That round hardened
the one class that was failing at the time; the identical exception then surfaced
in the next class to touch the same files. So the retry is now shared, and owned by
the RESOURCE rather than by whichever class is currently unlucky.

Correcting the premise on the way, because it changes what a fix can be: a
NotInParallel key cannot help here. CI runs this suite with
--maximum-parallel-tests 1, so there is no concurrency to serialise — meaning a
lock at Before(Test) time proves a handle OUTLIVED its owning test, either an
undisposed stream awaiting finalization or a child process not yet reaped. Adding
a shared key would have looked like a fix while the leak persisted, which is worse
than the flake.

The helper keeps the earlier conclusion intact: retry a transient sharing
violation, then THROW with a named cause. Never swallow — these tests assert on
token and profile state, and a stale tokens directory already satisfies "a peer
already refreshed it", so a false pass would hide a real regression rather than
cost a rerun.

One addition: a single GC + finalizer pass after the first failed attempt. It can
only release an in-process undisposed stream, so it doubles as a discriminator — if
the retry budget stops being exhausted, the holder was an unreferenced stream; if
failures continue, it is a live holder and the owner is still running. The thrown
message says the pass was attempted so the next occurrence carries that fact.

NOT verifiable locally: this is Windows-only (Unix unlinks regardless of open
handles) and intermittent, so neither a local run nor a single green Windows CI run
establishes it. Verified only that behaviour is unchanged where it can be:
TokenStoreProfileTests 20/20, CrossProcessRefreshTests 4/4.
