+++ test/Capacitor.Cli.Tests.Unit/TokenStoreProfileTests.cs
@@ -22,18 +22,13 @@ public class TokenStoreProfileTests {
 
     [Before(Test)]
     public void Cleanup() {
-        if (File.Exists(LegacyPath)) File.Delete(LegacyPath);
-        if (Directory.Exists(TokensDir)) Directory.Delete(TokensDir, recursive: true);
-
-        // Reset the shared profile config so the active profile resolves to "default".
-        // A config.json left in the shared KCAP_CONFIG_DIR by another test would make
-        // LoadAsync() resolve a different (file-less) profile, turning the legacy-fallback
-        // assertions order-dependent.
-        var cfg = Capacitor.Cli.Core.Config.AppConfig.GetConfigPath();
-        if (File.Exists(cfg)) File.Delete(cfg);
-        // Token lookup now consults AppConfig.ResolvedProfile, so a value left behind by another
-        // test would redirect these reads to a different profile.
-        Capacitor.Cli.Core.Config.AppConfig.ResetResolvedStateForTesting();
+        // Bare File.Delete/Directory.Delete here is what made this hook fail on Windows: the shared
+        // KCAP_CONFIG_DIR artefacts can still be held when this runs, and Windows treats that as a
+        // hard sharing violation, which from a Before hook fails the test before it starts. The
+        // shared helper retries the transient window and then throws with a named cause rather than
+        // running against another test's state. See SharedConfigDirCleanup for why this belongs to
+        // the resource rather than to whichever class happens to be failing.
+        SharedConfigDirCleanup.ClearTokenAndProfileState(LegacyPath, TokensDir);
     }
 
     [Test]
